backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
paho-mqtt==1.6.1